# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: StayHub V2 API
  version: 1.0.0
  description: Static OpenAPI contract aligned with the currently implemented StayHub V2 controllers.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v2/accommodations/{accommodationId}/reviews:
    post:
      tags:
        - Reviews
      summary: Create a review for a completed stay
      description: Allows the guest who completed a reservation to create a single review for the related accommodation.
      operationId: createReview
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: accommodationId
          required: true
          description: Accommodation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Review policy violation or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized, valid JWT token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only the guest owner can create the review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Accommodation or reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Reviews
      summary: List reviews for an accommodation
      description: Returns all reviews registered for the specified accommodation ordered from newest to oldest.
      operationId: getAccommodationReviews
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: accommodationId
          required: true
          description: Accommodation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
        '401':
          description: Unauthorized, valid JWT token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Accommodation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v2/reviews/{reviewId}/response:
    post:
      tags:
        - Reviews
      summary: Answer a guest review
      description: Allows the host owner of the accommodation to respond once to a guest review.
      operationId: respondToReview
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reviewId
          required: true
          description: Review identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondReviewRequest'
      responses:
        '200':
          description: Review answered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Review policy violation or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized, valid JWT token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only the host owner can answer the review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      description: Standard error payload returned by the API.
      properties:
        message:
          type: string
          example: Invalid login attempt detected
        code:
          type: integer
          example: 401
      required:
        - message
        - code
...
    CreateReviewRequest:
      type: object
      properties:
        reservationId:
          type: integer
          format: int64
          minimum: 1
          example: 45
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          maxLength: 1000
          example: Excelente alojamiento, muy limpio y bien ubicado.
      required:
        - reservationId
        - rating
        - comment
    RespondReviewRequest:
      type: object
      properties:
        response:
          type: string
          maxLength: 1000
          example: Gracias por tu visita. Nos alegra saber que disfrutaste la estadia.
      required:
        - response
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12
        accommodationId:
          type: integer
          format: int64
          example: 7
        guestId:
          type: integer
          format: int64
          example: 18
        guestName:
          type: string
          example: Juan Perez
        stayStartDate:
          type: string
          format: date-time
          example: '2026-05-10T15:00:00'
        stayEndDate:
          type: string
          format: date-time
          example: '2026-05-14T11:00:00'
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Excelente alojamiento, muy limpio y con una vista hermosa.
        hostResponse:
          type: string
          nullable: true
          example: Muchas gracias por tu visita. Siempre seras bienvenido.
        hostRespondedAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-16T09:30:00'
        createdAt:
          type: string
          format: date-time
          example: '2026-05-15T18:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-05-16T09:30:00'
      required:
        - id
        - accommodationId
        - guestId
        - guestName
        - stayStartDate
        - stayEndDate
        - rating
        - comment
        - createdAt
        - updatedAt
...